requirements.txt
pandas==2.1.3
openpyxl==3.1.2
slack-sdk==3.26.1
python-dotenv==1.0.0
requests==2.32.3 